+++ client-gwt/src/test/java/org/realityforge/replicant/client/gwt/TestClientSession.java
@@ -1,0 +1,21 @@
+package org.realityforge.replicant.client.gwt;
+
+import javax.annotation.Nonnull;
+import org.realityforge.replicant.client.transport.AbstractDataLoaderService;
+import org.realityforge.replicant.client.transport.ClientSession;
+import static org.mockito.Mockito.*;
+
+public class TestClientSession
+  extends ClientSession<TestClientSession, TestGraph>
+{
+  public TestClientSession( @Nonnull final AbstractDataLoaderService<TestClientSession, TestGraph> dataLoaderService,
+                            @Nonnull final String sessionID )
+  {
+    super( dataLoaderService, sessionID );
+  }
+
+  public TestClientSession( @Nonnull final String sessionID )
+  {
+    this( mock( AbstractDataLoaderService.class ), sessionID );
+  }
+}